requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django
django-admin-sortable2
django-cors-headers==4.4.0
djangorestframework
idna==3.10
pytz
requests
sqlparse==0.5.1
urllib3==2.2.3
python-json-logger==2.0.7
tqdm